Move interfaces into separate project
