+++ Services/Webstore.Interfaces/Services/ICartService.cs
@@ -1,6 +1,6 @@
 ﻿using WebStore.Domain.ViewModels;
 
-namespace WebStore.Infrastructure.Interfaces
+namespace Webstore.Interfaces.Services
 {
     public interface ICartService
     {
