+++ Services/Webstore.Interfaces/Services/IEmployeesData.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using WebStore.Domain.Models;
 
-namespace WebStore.Infrastructure.Interfaces
+namespace Webstore.Interfaces.Services
 {
     public interface IEmployeesData
     {
