+++ UI/WebStore/Components/BrandsViewComponent.cs
@@ -3,7 +3,7 @@ using System.Linq;
 
 using Microsoft.AspNetCore.Mvc;
 using WebStore.Domain.ViewModels;
-using WebStore.Infrastructure.Interfaces;
+using Webstore.Interfaces.Services;
 
 namespace WebStore.Components
 {
