+++ UI/WebStore/Controllers/CatalogController.cs
@@ -2,8 +2,8 @@
 using Microsoft.AspNetCore.Mvc;
 using WebStore.Domain;
 using WebStore.Domain.ViewModels;
-using WebStore.Infrastructure.Interfaces;
 using WebStore.Infrastructure.Mapping;
+using Webstore.Interfaces.Services;
 
 namespace WebStore.Controllers
 {
