+++ UI/WebStore/Infrastructure/Services/InCookies/InCookiesCartService.cs
@@ -9,8 +9,8 @@ using Newtonsoft.Json;
 using WebStore.Domain;
 using WebStore.Domain.Entities;
 using WebStore.Domain.ViewModels;
-using WebStore.Infrastructure.Interfaces;
 using WebStore.Infrastructure.Mapping;
+using Webstore.Interfaces.Services;
 
 namespace WebStore.Infrastructure.Services.InCookies
 {
