+++ UI/WebStore/Infrastructure/Services/InMemory/InMemoryProductData.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using WebStore.Data;
 using WebStore.Domain;
 using WebStore.Domain.Entities;
-using WebStore.Infrastructure.Interfaces;
+using Webstore.Interfaces.Services;
 
 namespace WebStore.Infrastructure.Services.InMemory
 {
