+++ UI/WebStore/Infrastructure/Services/InSQL/SqlOrderService.cs
@@ -8,7 +8,7 @@ using WebStore.DAL.Context;
 using WebStore.Domain.Entities.Identity;
 using WebStore.Domain.Entities.Orders;
 using WebStore.Domain.ViewModels;
-using WebStore.Infrastructure.Interfaces;
+using Webstore.Interfaces.Services;
 
 namespace WebStore.Infrastructure.Services.InSQL
 {
