+++ UI/WebStore/Infrastructure/Services/InSQL/SqlProductData.cs
@@ -6,7 +6,7 @@ using Microsoft.EntityFrameworkCore;
 using WebStore.DAL.Context;
 using WebStore.Domain;
 using WebStore.Domain.Entities;
-using WebStore.Infrastructure.Interfaces;
+using Webstore.Interfaces.Services;
 
 namespace WebStore.Infrastructure.Services.InSQL
 {
