+++ UI/WebStore/Views/Home/Index.cshtml
@@ -1,5 +1,5 @@
-﻿@using WebStore.Infrastructure.Interfaces
-@using WebStore.Domain.ViewModels
+﻿@using WebStore.Domain.ViewModels
+@using Webstore.Interfaces.Services
 @inject IProductData ProductData
 @{
     var products = ProductData
