+++ UI/WebStore/WebStore.csproj
@@ -21,6 +21,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Common\WebStore.Domain\WebStore.Domain.csproj" />
     <ProjectReference Include="..\..\Services\WebStore.DAL\WebStore.DAL.csproj" />
+    <ProjectReference Include="..\..\Services\Webstore.Interfaces\Webstore.Interfaces.csproj" />
   </ItemGroup>
 
   <ItemGroup>
